adds Alert and Breadcrumbs from DaisyUi
Alert is set via useContext so that is available everywhere. this allows for messages
to be set for all kind of events such as successful purchase, removal of items from
cart, or no tikets available. Breadcrumbs have been added in Event and Attraction
page, these have links that will bring the user back on click.