simplyfies AttractionShowIcons and BackArrowOverlay
AttractionShowIcons now uses two arrays, one with the name of the data.ExternalLinks
name that is lookign for, one for the react icon that needs to be used at that index. it uses map to iterate to the names array and check if it exist. in that case it appends to the result the icon with the tailwind styling. previously this was done imperatively. other update is BackArrowOverlay, not the icon to go back in the page is contained in his own fucntion instead of being straigth in the return statement