add darkmode
using tailwindcss and a daisyUi toggle component the page switch between dark and
light mode. taiwind.config adds darkmode:class to enable the use of dark: when this
class is active globally in the body. the toggle element removes or adds "dark" or
"light" class in the root. useEffect mantains the style when changing pages