add duration-[1000ms] to darkmode and useOutsideAlerter changed
the 1sec duration allows for the colors to blend while changing from light to dark
mode the useOutsideAlerter to close the Cart has been changed to allow the darkmode
toggle to be clicked without closing the cart. it will save ther eference of the
toggle component then compare the event.target with it. if ref, the ref.current doesnt contain event.target and the event.target is not the toggle element, then close the cart.