add happy_cart icon when checkout is clicked
the sad-cart icon is modified to have a smile when the user click on the checkout
button, to indicate successful purchase. the icon revert to sad when at least an items
is added to the cart and then removed (as it wont show unless the cart is completely
empty, so when i buy something it will keep the happy smile, if i add something the
icon wont be showed, if i purchase the icon is still happy, but if i remove it change
to sad). a state isCheckoutClicked is added to cart and is set to true on checkout
pressed. alert for emty basket has been removed, as there is no checkout button when
empty.