add dynamic messages to AddToCart
a state is introduced, to keep a note of the message to be displayed on the button, if
the item on screen is already on the cart, and if the item is available. an UseEffect
does cart.item.some() to check if the id of the item we are lookign matches the id of
one of the item in the cart, if so, it will set the message for ALREADY IN THE CART,
and isInCArt:true. it will also look for not added (ADD TO CART) and unavailable (NO
TICKETS AVAILABLE)

onClick for the button has been replaced with his own function, as now we have to show
another alert if is altready in the basket and the ternary operator wasnt feasible