solve re-render of entire page when clicking cart
the state that manage the display of the cart has been moved inside the
cart instead of being in a state in the main.js. this doesnt force the
entire page to render everytime cart state who was on the main file is
changed, as now the state is stored in the navbar. Suspense for metadata
and map have been restored in their original position as it seems that
the issue is resolved for now.