refactor ArtistComponents, opens cart when AddToCart is clicked
ArtistComponent has been refactored to have different files and be
easier to read

AddToCart had previously his ability to open the cart disabled due to
the fact that the cart.display wasn't used anymore to open the cart;
instead it was using a state inside the navbar to avoid render the main
everytime i was opening the cart. instead of putting the state up again or usecontext (which was going to create the same issue) i decided to manipulate the DOM directly. when AddToCart is clicked, it changes the style.display to "block"