BUGFIX: searchbar & results + metadata not updating
the results of the searchbar were offcenter due to the absolute position being set to auto and the navbar having to
share space with the darkmode button. the latter has been moved away. previously the metadata and top5 weren't
updating when settign anew artist via the searchbar. this has been fixed by having the fetch hook to conditionally
execute when the dependency array changes